+++ src/snowflake/snowpark/dataframe.py
@@ -3323,7 +3323,7 @@ class DataFrame:
     @property
     def session(self) -> "snowflake.snowpark.Session":
         """
-        Returns a :class:`snowflake.snowpark.Session` object that provides access to the session current dataframe is relying on.
+        Returns a :class:`snowflake.snowpark.Session` object that provides access to the session the current DataFrame is relying on.
         """
         return self._session