+++ src/snowflake/snowpark/session.py
@@ -1683,14 +1683,14 @@ class Session:
 
     @property
     def session_id(self) -> int:
-        """Returns an integer that represent the session id of this session."""
+        """Returns an integer that represents the session id of this session."""
         return self._session_id
 
     @property
-    def connection(self) -> "ServerConnection":
-        """Returns a :class:`ServerConnection` object that allow you to access the connection between current session
+    def connection(self) -> "SnowflakeConnection":
+        """Returns a :class:`SnowflakeConnection` object that allow you to access the connection between current session
         and snowflake server"""
-        return self._conn
+        return self._conn._conn
 
 
     def _run_query(